sign out option solved
