+++ app/src/main/java/au/mgemployeehire/employeehire/SubmissionSummeryActivity.java
@@ -18,7 +18,7 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
             suburbTV, stateTV, nameOfThePersonTV, phoneTV, ppeTV, transportRequirementsTV,
             engRequirementTV, liftingCapacityTV, additionalRequirementTV, licenseRequiredTV, environmentTV, emailTV,
             supervisorMobileNumberTV, yourNameTV, workSiteStreetTV, workSiteSuburbTV,
-            workSiteStateTV, workerQuantityTV, workingDivisionTV, startTimeTV, endTimeTV, maleFemaleTV ;//26 variables
+            workSiteStateTV, workerQuantityTV, workingDivisionTV, startTimeTV, endTimeTV, maleFemaleTV, dearJohnTV, awardTV ;//26 variables
 
     private Button confirmButton, backButton;
     DatabaseReference databaseReference;
@@ -62,6 +62,8 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
         startTimeTV = findViewById(R.id.startTimeTVID);
         endTimeTV = findViewById(R.id.endTimeTVID);
         maleFemaleTV = findViewById(R.id.maleFemaleTVID);
+        dearJohnTV = findViewById(R.id.dearJohnTVID);
+        awardTV = findViewById(R.id.awardTVID);
 
         //getting the values
         //getting from date
@@ -210,6 +212,7 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
         if (yourName!=null){
             String yourNameStr = yourName.getString("yourName");
             yourNameTV.setText(yourNameStr);
+            dearJohnTV.setText(yourNameStr);
         }
 
         //work site location
@@ -269,6 +272,13 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
             maleFemaleTV.setText(maleFemaleStr);
         }
 
+        //award
+        Bundle awardBundle = getIntent().getExtras();
+        if (awardBundle!=null){
+            String awardStr = awardBundle.getString("award");
+            awardTV.setText(awardStr);
+        }
+
 
         //functioning confirm button. Sending data to firebase Database
         confirmButton.setOnClickListener(new View.OnClickListener() {
@@ -320,6 +330,7 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
         String workingDivisionStr = workingDivisionTV.getText().toString().trim();
         String startTimeStr = startTimeTV.getText().toString().trim();
         String endTimeStr = endTimeTV.getText().toString().trim();
+        String awardStr = awardTV.getText().toString().trim();
 
         //key generate
         String key = databaseReference.push().getKey();
@@ -352,7 +363,8 @@ public class SubmissionSummeryActivity extends AppCompatActivity{
                 liftingCapacityStr,
                 additionalRequirementStr,
                 environmentStr,
-                licenseRequiredStr
+                licenseRequiredStr,
+                awardStr
                 );
 
         //sending job advertisement data to database
