+++ app/src/main/res/layout/activity_job_position_selection.xml
@@ -83,23 +83,21 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
+        android:layout_marginTop="10dp"
         android:orientation="horizontal"
         android:gravity="center">
 
         <Button
             android:id="@+id/backJobPositionSelectionPageBtnID"
-            android:layout_width="150dp"
+            android:layout_width="120dp"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="20dp"
             android:text="Back"
             app:backgroundTint="#00BCD4" />
 
         <Button
             android:id="@+id/JobPositionSelectionNextBtnID"
-            android:layout_width="150dp"
+            android:layout_width="120dp"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="20dp"
             android:text="Next"
             android:layout_marginLeft="5dp"
             app:backgroundTint="#00BCD4" />
